+++ src/qobj/operators.jl
@@ -37,11 +37,11 @@ rand_unitary(::Type{T}, dimensions::Int, distribution::Union{Symbol, Val} = Val(
     rand_unitary(T, SVector(dimensions), makeVal(distribution); rng)
 rand_unitary(
     ::Type{T},
-    dimensions::Union{Dimensions, VectorOrTuple{Int}},
+    dimensions::Union{Dimensions, AbstractVecOrTuple{Int}},
     distribution::Union{Symbol, Val} = Val(:haar);
     rng::AbstractRNG = default_rng(),
 ) where {T <: FloatOrComplex} = rand_unitary(T, dimensions, makeVal(distribution); rng)
-function rand_unitary(::Type{T}, dimensions::Union{Dimensions, VectorOrTuple{Int}}, ::Val{:haar}; rng::AbstractRNG = default_rng()) where {T <: FloatOrComplex}
+function rand_unitary(::Type{T}, dimensions::Union{Dimensions, AbstractVecOrTuple{Int}}, ::Val{:haar}; rng::AbstractRNG = default_rng()) where {T <: FloatOrComplex}
     N = get_size(dimensions)[1]
 
     # generate N x N matrix Z of complex standard normal random variates
@@ -58,7 +58,7 @@ function rand_unitary(::Type{T}, dimensions::Union{Dimensions, VectorOrTuple{Int
 end
 function rand_unitary(
         ::Type{T},
-        dimensions::Union{Dimensions, VectorOrTuple{Int}},
+        dimensions::Union{Dimensions, AbstractVecOrTuple{Int}},
         ::Val{:exp};
         rng::AbstractRNG = default_rng()
     ) where {T <: FloatOrComplex}
@@ -72,9 +72,9 @@ function rand_unitary(
     H = QuantumObject((Z + Z') / 2; type = Operator(), dims = dimensions)
     return to_dense(_rand_unitary_exp(H))
 end
-rand_unitary(::Type{T}, dimensions::Union{Dimensions, VectorOrTuple{Int}}, ::Val{Td}; rng::AbstractRNG = default_rng()) where {T <: FloatOrComplex, Td} =
+rand_unitary(::Type{T}, dimensions::Union{Dimensions, AbstractVecOrTuple{Int}}, ::Val{Td}; rng::AbstractRNG = default_rng()) where {T <: FloatOrComplex, Td} =
     throw(ArgumentError("Invalid distribution: $(Td)"))
-rand_unitary(dimensions::Union{Int, Dimensions, VectorOrTuple{Int}}, distribution::Union{Symbol, Val} = Val(:haar); rng::AbstractRNG = default_rng()) =
+rand_unitary(dimensions::Union{Int, Dimensions, AbstractVecOrTuple{Int}}, distribution::Union{Symbol, Val} = Val(:haar); rng::AbstractRNG = default_rng()) =
     rand_unitary(ComplexF64, dimensions, distribution; rng)
 
 # we make H sparse here because the following method currently does not exist : exp(::Matrix{Complex{BigFloat}})
@@ -593,9 +593,9 @@ where ``\omega = \exp(\frac{2 \pi i}{N})``.
     It is highly recommended to use `qft(dimensions)` with `dimensions` as `Tuple` or `SVector` from [StaticArrays.jl](https://github.com/JuliaArrays/StaticArrays.jl) to keep type stability. See the [related Section](@ref doc:Type-Stability) about type stability for more details.
 """
 qft(::Type{T}, dimensions::Int) where {T <: Complex} = QuantumObject(_qft_op(T, dimensions), Operator(), dimensions)
-qft(::Type{T}, dimensions::Union{Dimensions, VectorOrTuple{Int}}) where {T <: Complex} =
+qft(::Type{T}, dimensions::Union{Dimensions, AbstractVecOrTuple{Int}}) where {T <: Complex} =
     QuantumObject(_qft_op(T, get_size(dimensions)[1]), Operator(), dimensions)
-qft(dimensions::Union{Int, Dimensions, VectorOrTuple{Int}}) = qft(ComplexF64, dimensions)
+qft(dimensions::Union{Int, Dimensions, AbstractVecOrTuple{Int}}) = qft(ComplexF64, dimensions)
 
 function _qft_op(::Type{T}, N::Int) where {T <: Complex}
     N_T = T(N)